refactored, combined and cleaned up critical path functionalities
